document_engine/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "document-engine"
version = "0.1.0"
description = "Deterministic document completeness engine"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pdfminer.six>=20231228",
  "rapidfuzz>=3.9.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "paddleocr>=2.8.1",
  "opencv-python-headless>=4.10.0",
  "redis>=5.0.0",
  "rq>=1.16.0",
  "psycopg[binary]>=3.1.19"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
